added guard to prevent specs from blowing up on machines without bunny
I recognize that we are being a little biased here WRT mandating some
adapter dependencies and not others.  We will need to come up with a
better system of handling the different adapters so that contributors
can just run the specs for the adapters they are interested in and just
get friendly warnings for the other ones.  Oh, and we need to set up a
CI server to test all of the adapters. :)